move navbar into routemap.js and remove initialAppState.js
There is no longer any need for the initialAppState.js JavaScript file,
as the only thing left in it was the navbar, which probably belongs
with the other routing information in routemap.js
