--- frontend/@types/@dashlive/init.d.ts
@@ -1,3 +1,0 @@
-import { NavBarItem } from "../../src/types/NavBarItem";
-
-export const navbar: NavBarItem[];
