+++ frontend/@types/@dashlive/routemap.d.ts
@@ -1,5 +1,8 @@
-import { RouteMap } from "../../src/types/RouteMap";
+import type { RouteMap } from "../../src/types/RouteMap";
+import type { NavBarItem } from "../../src/types/NavBarItem";
 
 export const routeMap: RouteMap;
 
 export const uiRouteMap: RouteMap;
+
+export const navbar: NavBarItem[];
