+++ frontend/html/partials/importmap.hbs
@@ -1,6 +1,5 @@
 {
   "imports": {
-    "@dashlive/init": "/libs/initialAppState.js",
     "@dashlive/options": "/libs/options.js",
     "@dashlive/routemap": "/libs/routemap.js",
     "socket.io": "/libs/socket.io.js"
