--- frontend/src/test/fixtures/initialAppState.js
@@ -1,32 +1,0 @@
-export const navbar = [
-  {
-    active: true,
-    className: "spa",
-    href: "/",
-    title: "Home",
-  },
-  {
-    active: false,
-    className: "",
-    href: "/streams",
-    title: "Streams",
-  },
-  {
-    active: false,
-    className: "spa",
-    href: "/multi-period-streams",
-    title: "Multi-Period",
-  },
-  {
-    active: false,
-    className: "",
-    href: "/validate/",
-    title: "Validate",
-  },
-  {
-    active: false,
-    className: "",
-    href: "/media/inspect",
-    title: "Inspect",
-  }
-];
