--- templates/esm/initial_app_state.tjs
@@ -1,1 +1,0 @@
-export const navbar = {{ navbar | toJson }};
