+++ templates/esm/routemap.tjs
@@ -19,3 +19,5 @@ export const uiRouteMap = {
   },
 {%- endfor %}
 };
+
+export const navbar = {{ navbar | toJson }};
